+++ samples/page-scroller/page-scroll-view-dynamic-example.cpp
@@ -96,7 +96,7 @@ private:
   {
     Window window = application.GetWindow();
     auto positionSize = window.GetPositionSize();
-    window.SetPositionSize(Dali::PositionSize(positionSize.x, positionSize.y, WINDOW_W, WINDOW_H));
+    window.SetPositionSize(Dali::PositionSize(positionSize.x, positionSize.y, static_cast<int32_t>(WINDOW_W), static_cast<int32_t>(WINDOW_H)));
     window.SetBackgroundColor(Color::BLACK);
     window.KeyEventSignal().Connect(this, &PageScrollDynamicController::OnKeyEvent);
