+++ samples/text/text-input-field-example.cpp
@@ -201,7 +201,7 @@ private:
     scrollContent.SetSpacing(STACK_SPACING);
     scrollContent.SetRequestedWidth(MATCH_PARENT);
     scrollContent.SetRequestedHeight(WRAP_CONTENT);
-    scrollContent.SetPadding(Extents(0, 0, 0, STACK_PADDING));
+    scrollContent.SetPadding(Extents(0, 0, 0, static_cast<int16_t>(STACK_PADDING)));
     scrollContent.Add(cursorRow1);
     scrollContent.Add(cursorRow2);
     scrollContent.Add(placeholderRow);
@@ -227,7 +227,7 @@ private:
     rootLayout.SetSpacing(STACK_SPACING);
     rootLayout.SetRequestedWidth(MATCH_PARENT);
     rootLayout.SetRequestedHeight(MATCH_PARENT);
-    rootLayout.SetPadding(Extents(STACK_PADDING, STACK_PADDING, STACK_PADDING, STACK_PADDING));
+    rootLayout.SetPadding(Extents(static_cast<int16_t>(STACK_PADDING), static_cast<int16_t>(STACK_PADDING), static_cast<int16_t>(STACK_PADDING), static_cast<int16_t>(STACK_PADDING)));
     rootLayout.Add(fixedHeader);
     rootLayout.Add(scrollView);
